move unitary matrices enum out of qcp/cli to qcp/algorithms
